always jack's fault smh -jack

+++ xv6-riscv/user/schizo_threads.h
@@ -14,7 +14,7 @@ typedef void(*sthread_fn_in_t)(void *);
 int sthread_create(sthread_t *thread, sthread_fn_in_t fn, void *args, void *stack);
 
 
-int sthread_join(sthread_t *thread);
+uint64 sthread_join(sthread_t *thread);
 
 
 int sthread_exit();
